refactor units
